SwiftUIThinking ToDoList#6

+++ SwiftUITodoList/SwiftUITodoList/Views/ListView.swift
@@ -4,7 +4,7 @@
 //
 //  Created by sungyeon kim on 2022/12/28.
 //
-
+ 
 import SwiftUI
 
 struct ListView: View {
@@ -12,19 +12,27 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
 
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem(indexSet:))
+                    .onMove(perform: listViewModel.moveItem(from:to:))
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem(indexSet:))
-            .onMove(perform: listViewModel.moveItem(from:to:))
-        }
-        .listStyle(PlainListStyle())
+        }//: ZSTACK
+
         .navigationTitle("Todo List 🖥️")
         .navigationBarItems(
             leading: EditButton(),
